+++ src/frontend/pages/components/modalpage/modal.tsx
@@ -1,19 +1,18 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState, useContext } from "react";
+import { SocketContext } from "../../../context/socket";
 import "../../index.css";
 
 export default function ModalBasic({
   setIsOpenModal,
-  socket,
   roomname,
   innerText,
 }: {
   setIsOpenModal: any;
-  socket: any;
   roomname: string;
   innerText: string;
 }) {
   // 모달 끄기
-
+  const socket = useContext(SocketContext);
   const [passWord, setPassWord] = useState("");
   const [disabled, setDisabled] = useState(false);
