+++ src/frontend/pages/components/searchlistpage/ErrorMessage.tsx
@@ -1,0 +1,9 @@
+export default function ErrorMessage({ message }: { message: string }) {
+  console.log("errmessga called");
+  return (
+    <p className="error">
+      <span>📛</span>
+      {message}
+    </p>
+  );
+}
